orchestrator/pkg/orchestrator/service.py

import json
from google.cloud import pubsub_v1
from google.cloud import aiplatform
from pkg.config import cfg

# Initialize clients once to be reused across function invocations.
publisher = pubsub_v1.PublisherClient()
topic_path = publisher.topic_path(cfg.gcp_project_id, cfg.executor_topic_id)
aiplatform.init(project=cfg.gcp_project_id, location="us-central1")

def process_opportunity(data: dict):
    """
    Processes a potential arbitrage opportunity.
    1. Validates the opportunity with Vertex AI.
    2. If valid, enriches the data and publishes it to the executor topic.
    """
    print("Orchestrating opportunity...")

    # 1. Prepare the prompt for the Vertex AI Gemini model.
    # In a real scenario, you would include much more context:
    # real-time gas fees, liquidity depth from other APIs, historical data, etc.
    prompt = f"""
    Analyze the following potential arbitrage opportunity:
    - Strategy: {data.get('strategy')}
    - Pair: {data.get('pair')}
    - DEX A: {data.get('dexA')}
    - DEX B: {data.get('dexB')}
    - Potential Profit: {data.get('potentialProfit')}

    Based on a high-risk, high-reward profile, is this opportunity worth pursuing?
    Consider a maximum acceptable slippage of 0.1%.
    Respond with a JSON object containing: {{ "isValid": boolean, "reason": string, "requiredFlashLoan": float }}.
    """
    print(f"  - Sending prompt to Vertex AI model: {cfg.vertex_ai_model}")

    # 2. Simulate the call to Vertex AI.
    # In a real implementation, you would use the Vertex AI client library:
    # from vertexai.generative_models import GenerativeModel
    # model = GenerativeModel(cfg.vertex_ai_model)
    # response = model.generate_content(prompt)
    # ai_response = json.loads(response.text)
    
    # Mocked response for this blueprint
    ai_response = {
        "isValid": True,
        "reason": "High probability of success with low expected slippage.",
        "requiredFlashLoan": 50.0
    }
    print(f"  - Received validation from Vertex AI: {ai_response}")

    # 3. If the AI validates the opportunity, publish it to the executor.
    if ai_response.get("isValid"):
        # Enrich the data with the AI's validation and requirements.
        enriched_data = {**data, "aiValidation": ai_response}
        
        message_data = json.dumps(enriched_data).encode("utf-8")
        
        future = publisher.publish(topic_path, message_data)
        message_id = future.result()
        
        print(f"  - Published validated opportunity to executor topic; msg ID: {message_id}")
    else:
        print(f"  - Opportunity deemed not viable by Vertex AI. Reason: {ai_response.get('reason')}")

