orchestrator/requirements.txt

Flask==3.0.0
gunicorn==21.2.0
google-cloud-pubsub==2.19.0
google-cloud-aiplatform==1.46.0
